Time: 0 ms (100.00%), Space: 5.9 MB (71.30%) - LeetHub

+++ 1359-count-all-valid-pickup-and-delivery-options/1359-count-all-valid-pickup-and-delivery-options.cpp
@@ -3,7 +3,7 @@ public:
     int countOrders(int n) {
         int res = 1,mod = 1e9+7;
         for(int i =1;i<=n;i++){
-            res = ((((long long int)2*i-1)%mod)*(((long long int)i*res)%mod))%mod;
+            res = (((2*i-1)%mod)*(((long long int)i*res)%mod))%mod;
         }
         return res;
     }
